split emit file read host effects
