+++ docs/COMPLETION_AUDIT.md
@@ -308,6 +308,11 @@ and do not assume Phase 4 is ready without evidence.
   `tests/integration/cli_build/direct_build_graph_host_effects/file_reads.rs`,
   leaving env-effect ordering and unrelated test skip coverage in the parent
   module.
+- Emit-command build graph host-effect integration tests now split declared
+  file-read fallback accept/reject coverage into
+  `tests/integration/cli_build/emit_direct_host_effects/file_reads.rs`,
+  leaving env-effect ordering and unrelated test skip coverage in the parent
+  module.
 - Build graph unit tests now split target metadata validation and gated
   package/link target rejection coverage into
   `tests/build_graph/target_metadata.rs`, leaving the root build graph test
